Make name consistent
This property is called buildListenerWithArtifacts for listener and buildAndroidWithArtifacts for android so it seems reasonable to name it buildJavaWithArtifacts here. I actually don't understand why there are three different switches. I would think they should all have the same name since it seems weird that someone would build with artifacts in one place and not another.