CBL-583: Prevent localDoc object to be collected (#182)
It seems like localDoc could be collected before the end of saveResolvedDocument(). The speculation is that the JIT optimizer might think that localDoc hasn’t been used anymore after the line that gets the rawDoc (C4Document) from the localDoc so it could be collected.

From repeatedly runing the test, this little trick by trying to use the localDoc until the end could help keeping the localDoc alive until the end.